reducing API fetching number
